Action mailer with active job using sidekiq implemented

+++ app/controllers/audit_logs_controller.rb
@@ -2,17 +2,20 @@ class AuditLogsController < ApplicationController
     # before_action :authenticate_user!, :authorize_super_admin
     before_action :set_branch
     before_action :set_audit_log, only: [:show, :destroy]
+    after_action :verify_authorized
 
     
     def index
-        @audit_logs = @branch.audit_logs.includes(:cashier, :medicine, :record).order(created_at: :desc)
+        @audit_logs = policy_scope(AuditLog)
+        authorize @audit_logs
     end
 
     def show
-        @audit_log = AuditLog.find(params[:id])
+        authorize @audit_log
     end
     
     def destroy
+        authorize @audit_log
         if @audit_log.destroy
             redirect_to branch_audit_logs_path(@branch), notice: 'Audit log was successfully deleted.'
         else
